+++ app/src/main/java/com/topjohnwu/magisk/extensions/XAndroid.kt
@@ -1,5 +1,6 @@
 package com.topjohnwu.magisk.extensions
 
+import android.app.Activity
 import android.content.ComponentName
 import android.content.Context
 import android.content.ContextWrapper
@@ -22,17 +23,20 @@ import android.os.Build
 import android.os.Build.VERSION.SDK_INT
 import android.provider.OpenableColumns
 import android.view.View
+import android.view.inputmethod.InputMethodManager
 import androidx.annotation.ColorRes
 import androidx.annotation.DrawableRes
 import androidx.appcompat.content.res.AppCompatResources
 import androidx.core.content.ContextCompat
+import androidx.core.content.getSystemService
 import androidx.core.net.toFile
 import androidx.core.net.toUri
+import androidx.fragment.app.Fragment
 import com.topjohnwu.magisk.FileProvider
 import com.topjohnwu.magisk.core.Const
+import com.topjohnwu.magisk.core.utils.Utils
 import com.topjohnwu.magisk.core.utils.currentLocale
 import com.topjohnwu.magisk.utils.DynamicClassLoader
-import com.topjohnwu.magisk.core.utils.Utils
 import com.topjohnwu.superuser.Shell
 import com.topjohnwu.superuser.ShellUtils
 import java.io.File
@@ -347,3 +351,14 @@ val isAB
     get() = ShellUtils
         .fastCmd("grep_prop ro.build.ab_update")
         .let { it.isNotEmpty() && it.toBoolean() }
+
+fun Activity.hideKeyboard() {
+    val view = currentFocus ?: return
+    getSystemService<InputMethodManager>()
+        ?.hideSoftInputFromWindow(view.windowToken, 0)
+    view.clearFocus()
+}
+
+fun Fragment.hideKeyboard() {
+    activity?.hideKeyboard()
+}
