+++ app/src/main/java/com/topjohnwu/magisk/model/entity/internal/DownloadSubject.kt
@@ -2,13 +2,13 @@ package com.topjohnwu.magisk.model.entity.internal
 
 import android.content.Context
 import android.os.Parcelable
-import com.topjohnwu.magisk.core.Info
 import com.topjohnwu.magisk.core.Config
+import com.topjohnwu.magisk.core.Info
+import com.topjohnwu.magisk.core.model.MagiskJson
+import com.topjohnwu.magisk.core.model.ManagerJson
 import com.topjohnwu.magisk.core.model.module.Repo
 import com.topjohnwu.magisk.extensions.cachedFile
 import com.topjohnwu.magisk.extensions.get
-import com.topjohnwu.magisk.core.model.MagiskJson
-import com.topjohnwu.magisk.core.model.ManagerJson
 import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
 import java.io.File
