+++ app/src/main/java/com/topjohnwu/magisk/model/events/dialog/BiometricDialog.kt
@@ -1,9 +1,9 @@
 package com.topjohnwu.magisk.model.events.dialog
 
 import com.topjohnwu.magisk.core.base.BaseActivity
+import com.topjohnwu.magisk.core.utils.BiometricHelper
 import com.topjohnwu.magisk.model.events.ActivityExecutor
 import com.topjohnwu.magisk.model.events.ViewEvent
-import com.topjohnwu.magisk.core.utils.BiometricHelper
 
 class BiometricDialog(
     builder: Builder.() -> Unit
