+++ app/src/main/java/com/topjohnwu/magisk/model/events/dialog/ManagerInstallDialog.kt
@@ -1,7 +1,7 @@
 package com.topjohnwu.magisk.model.events.dialog
 
-import com.topjohnwu.magisk.core.Info
 import com.topjohnwu.magisk.R
+import com.topjohnwu.magisk.core.Info
 import com.topjohnwu.magisk.core.download.DownloadService
 import com.topjohnwu.magisk.extensions.res
 import com.topjohnwu.magisk.model.entity.internal.Configuration
