+++ app/src/main/java/com/topjohnwu/magisk/model/events/dialog/UninstallDialog.kt
@@ -1,12 +1,12 @@
 package com.topjohnwu.magisk.model.events.dialog
 
 import android.widget.Toast
-import com.topjohnwu.magisk.core.Info
 import com.topjohnwu.magisk.R
+import com.topjohnwu.magisk.core.Info
 import com.topjohnwu.magisk.core.download.DownloadService
+import com.topjohnwu.magisk.core.utils.Utils
 import com.topjohnwu.magisk.model.entity.internal.Configuration
 import com.topjohnwu.magisk.model.entity.internal.DownloadSubject
-import com.topjohnwu.magisk.core.utils.Utils
 import com.topjohnwu.magisk.view.MagiskDialog
 import com.topjohnwu.superuser.Shell
