+++ app/src/main/java/com/topjohnwu/magisk/ui/flash/FlashFragment.kt
@@ -2,10 +2,10 @@ package com.topjohnwu.magisk.ui.flash
 
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.databinding.FragmentFlashMd2Binding
-import com.topjohnwu.magisk.ui.base.CompatFragment
+import com.topjohnwu.magisk.ui.base.BaseUIFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class FlashFragment : CompatFragment<FlashViewModel, FragmentFlashMd2Binding>() {
+class FlashFragment : BaseUIFragment<FlashViewModel, FragmentFlashMd2Binding>() {
 
     override val layoutRes = R.layout.fragment_flash_md2
     override val viewModel by viewModel<FlashViewModel>()
