+++ app/src/main/java/com/topjohnwu/magisk/ui/hide/HideFragment.kt
@@ -11,12 +11,11 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.databinding.FragmentHideMd2Binding
-import com.topjohnwu.magisk.ui.base.CompatFragment
-import com.topjohnwu.magisk.ui.base.hideKeyboard
+import com.topjohnwu.magisk.ui.base.BaseUIFragment
 import com.topjohnwu.magisk.utils.MotionRevealHelper
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class HideFragment : CompatFragment<HideViewModel, FragmentHideMd2Binding>() {
+class HideFragment : BaseUIFragment<HideViewModel, FragmentHideMd2Binding>() {
 
     override val layoutRes = R.layout.fragment_hide_md2
     override val viewModel by viewModel<HideViewModel>()
