+++ app/src/main/java/com/topjohnwu/magisk/ui/home/HomeFragment.kt
@@ -6,10 +6,10 @@ import android.view.MenuItem
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.databinding.FragmentHomeMd2Binding
 import com.topjohnwu.magisk.model.navigation.Navigation
-import com.topjohnwu.magisk.ui.base.CompatFragment
+import com.topjohnwu.magisk.ui.base.BaseUIFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class HomeFragment : CompatFragment<HomeViewModel, FragmentHomeMd2Binding>() {
+class HomeFragment : BaseUIFragment<HomeViewModel, FragmentHomeMd2Binding>() {
 
     override val layoutRes = R.layout.fragment_home_md2
     override val viewModel by viewModel<HomeViewModel>()
