+++ app/src/main/java/com/topjohnwu/magisk/ui/install/InstallFragment.kt
@@ -4,10 +4,10 @@ import android.content.Intent
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.databinding.FragmentInstallMd2Binding
 import com.topjohnwu.magisk.model.events.RequestFileEvent
-import com.topjohnwu.magisk.ui.base.CompatFragment
+import com.topjohnwu.magisk.ui.base.BaseUIFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class InstallFragment : CompatFragment<InstallViewModel, FragmentInstallMd2Binding>() {
+class InstallFragment : BaseUIFragment<InstallViewModel, FragmentInstallMd2Binding>() {
 
     override val layoutRes = R.layout.fragment_install_md2
     override val viewModel by viewModel<InstallViewModel>()
