+++ app/src/main/java/com/topjohnwu/magisk/ui/safetynet/SafetynetViewModel.kt
@@ -3,6 +3,7 @@ package com.topjohnwu.magisk.ui.safetynet
 import androidx.databinding.Bindable
 import com.topjohnwu.magisk.BR
 import com.topjohnwu.magisk.R
+import com.topjohnwu.magisk.core.utils.SafetyNetHelper
 import com.topjohnwu.magisk.extensions.subscribeK
 import com.topjohnwu.magisk.model.events.SafetyNetResult
 import com.topjohnwu.magisk.model.events.UpdateSafetyNetEvent
@@ -10,7 +11,6 @@ import com.topjohnwu.magisk.ui.base.BaseViewModel
 import com.topjohnwu.magisk.ui.safetynet.SafetyNetState.*
 import com.topjohnwu.magisk.utils.KObservableField
 import com.topjohnwu.magisk.utils.RxBus
-import com.topjohnwu.magisk.core.utils.SafetyNetHelper
 
 enum class SafetyNetState {
     LOADING, PASS, FAILED, IDLE
