+++ app/src/main/java/com/topjohnwu/magisk/ui/settings/SettingsFragment.kt
@@ -4,11 +4,11 @@ import android.os.Bundle
 import android.view.View
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.databinding.FragmentSettingsMd2Binding
-import com.topjohnwu.magisk.ui.base.CompatFragment
+import com.topjohnwu.magisk.ui.base.BaseUIFragment
 import com.topjohnwu.magisk.utils.PinchZoomTouchListener
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class SettingsFragment : CompatFragment<SettingsViewModel, FragmentSettingsMd2Binding>() {
+class SettingsFragment : BaseUIFragment<SettingsViewModel, FragmentSettingsMd2Binding>() {
 
     override val layoutRes = R.layout.fragment_settings_md2
     override val viewModel by viewModel<SettingsViewModel>()
