+++ app/src/main/java/com/topjohnwu/magisk/ui/theme/ThemeFragment.kt
@@ -2,10 +2,10 @@ package com.topjohnwu.magisk.ui.theme
 
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.databinding.FragmentThemeMd2Binding
-import com.topjohnwu.magisk.ui.base.CompatFragment
+import com.topjohnwu.magisk.ui.base.BaseUIFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class ThemeFragment : CompatFragment<ThemeViewModel, FragmentThemeMd2Binding>() {
+class ThemeFragment : BaseUIFragment<ThemeViewModel, FragmentThemeMd2Binding>() {
 
     override val layoutRes = R.layout.fragment_theme_md2
     override val viewModel by viewModel<ThemeViewModel>()
